Revert the change to fix duplicate responses on Linux
Linux seems to have a bug where resolving an invalid domain hangs
until the socket times out.  This doesn't appear to affect other
operating systems.  The timeouts for invalid domains are definitely
worse than improper handling of duplicate responses, so revert the
code on Linux but keep it on other operating systems that do not
have the timeout issue.