Implement dohpath SvcParam [draft-ietf-add-svcb-dns-08]
This patch implements "dohpath" SvcParam proposed in
[draft-ietf-add-svcb-dns-08]. This parameter specifies a URI template
for the :path used in DNS-over-HTTPS requests.

"dohpath" is employed by [DDR], also a to-be-published Proposed Standard
that specifies how to upgrade DNS transport to a more secure one, i.d.,
DNS-over-TLS or DNS-over-HTTPS. DDR is deployed in the public DNS
resolvers including Cloudflare DNS, Google Public DNS, and Quad9.

[draft-ietf-add-svcb-dns-08]: https://datatracker.ietf.org/doc/draft-ietf-add-svcb-dns/08/
[DDR]: https://datatracker.ietf.org/doc/draft-ietf-add-ddr/